+++ NEWS.md
@@ -1,5 +1,12 @@
 # rsyn release notes
 
+## 0.0.2 (NOT RELEASED YET)
+
+Various API changes, including:
+
+* Transfer operations return a new `Summary` object including counters of how
+  much work was done, and of non-fatal errors.
+
 ## 0.0.1 (2020-05-13)
 
 Features: